refactor(structure): reorganized project into core/cli/math modules - Moved arithmetic/trigo/stats to src/math/ and include/math/ - Moved history to src/cli/ and include/cli/ - Added stub files for parser.c, types.c and their headers - Updated all #include paths to reflect new structure - Fixed numscanned uninitialized warning - Updated Makefile SRC list and build output paths

+++ include/math/trigo.h
@@ -8,6 +8,6 @@
 #define M_PI 3.14159
 #endif
 
-double trigo(float x, int r); 		 // Trigonometric functions
+double trigo(double x, int r); 		 // Trigonometric functions
 
 #endif
